# server/openapi.yaml
openapi: 3.0.0
info:
  title: Mini Supermarket API
  version: 1.0.0
  description: API cho hệ thống quản lý siêu thị mini
servers:
  - url: 'http://localhost:5000'
...
paths:
...
  /api/carts/:
    get:
      tags:
        - cart
      summary: get All Carts
      operationId: getAllCarts
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/customers:
    get:
      tags:
        - customer
      summary: Get all customers with pagination and filtering
      operationId: getAllCustomers
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-indexed)
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Records per page
        - in: query
          name: status
          schema:
            type: string
            enum:
              - active
              - inactive
              - pending
          description: Customer status filter
      responses:
        '200':
          description: Successfully retrieved customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
                  data:
                    type: array
        '400':
          description: Invalid query parameters
    post:
      tags:
        - customer
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: Nguyen Van A
                email:
                  type: string
                  format: email
                  example: customer@example.com
                phone:
                  type: string
                  pattern: '^[0-9]{10}$'
                  example: 0912345678
                address:
                  type: string
                  minLength: 5
                  maxLength: 500
                  example: 123 Main St
                membershipType:
                  type: string
                  enum:
                    - Standard
                    - Gold
                    - Silver
                  example: Standard
                age:
                  type: integer
                  minimum: 18
                  maximum: 120
                  example: 30
                status:
                  type: string
                  enum:
                    - active
                    - inactive
                    - pending
                  default: active
              required:
                - name
                - email
                - phone
      responses:
        '201':
          description: Customer created successfully
        '400':
          description: Invalid customer data
        '422':
          description: Validation error
...
  /api/feedbacks/:
    post:
      tags:
        - feedback
      summary: create Feedback
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - category
                - subject
                - customer_id
              properties:
                category:
                  type: string
                  enum:
                    - complaint
                    - suggestion
                    - praise
                subject:
                  type: string
                  example: Product quality issue
                detail:
                  type: string
                  example: The product arrived damaged and the packaging was open.
                customer_id:
                  type: string
                  pattern: '^[0-9a-fA-F]{24}$'
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                sentiment:
                  type: string
                  example: negative
                order_id:
                  type: string
                  pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
    get:
      tags:
        - feedback
      summary: get All Feedbacks
      operationId: getAllFeedbacks
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/invoices/:
    get:
      tags:
        - invoice
      summary: get All Invoices
      operationId: getAllInvoices
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    post:
      tags:
        - invoice
      summary: create Invoice
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/orders/:
    get:
      tags:
        - order
      summary: get All Orders
      operationId: getAllOrders
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    post:
      tags:
        - order
      summary: create Order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/products:
    get:
      tags:
        - product
      summary: Get all products with pagination and filtering
      operationId: getAllProducts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: status
          schema:
            type: string
            enum:
              - active
              - inactive
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
        '400':
          description: Invalid query parameters
...
  /api/products/:
    post:
      tags:
        - product
      summary: create Product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/promotions/:
    get:
      tags:
        - promotion
      summary: get All Promotions
      operationId: getAllPromotions
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    post:
      tags:
        - promotion
      summary: create Promotion
      operationId: createPromotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/sections/:
    get:
      tags:
        - section
      summary: get All Sections
      operationId: getAllSections
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    post:
      tags:
        - section
      summary: create Section
      operationId: createSection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/staff/:
    get:
      tags:
        - staff
      summary: get All Staff
      operationId: getAllStaff
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    post:
      tags:
        - staff
      summary: create Staff
      operationId: createStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/suppliers/:
    get:
      tags:
        - supplier
      summary: get All Suppliers
      operationId: getAllSuppliers
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    post:
      tags:
        - supplier
      summary: create Supplier
      operationId: createSupplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
...